Adding controls to the API and some tests.

+++ src/main/java/cfg/samples/ClaszAPI.java
@@ -33,6 +33,9 @@ public class ClaszAPI {
 	@GetMapping("/{code}")
 	public ResponseEntity<Clasz> findById(@PathVariable String code) {
 		Optional<Clasz> clasz = claszService.findById(code);
+		if (!clasz.isPresent()) {
+			return ResponseEntity.notFound().build();
+		}
 		return ResponseEntity.ok(clasz.get());
 	}
 	
@@ -43,12 +46,20 @@ public class ClaszAPI {
 	
 	@PutMapping("/{code}")
 	public ResponseEntity<Clasz> update(@PathVariable String code, @Valid @RequestBody Clasz clasz) {
+		Optional<Clasz> cl = claszService.findById(code);
+		if (!cl.isPresent() ) {
+			return ResponseEntity.notFound().build();
+		}
 		return ResponseEntity.ok(claszService.save(clasz));
 	}
 	
 	@SuppressWarnings("rawtypes")
 	@DeleteMapping("/{code}")
 	public ResponseEntity delete(@PathVariable String code) {
+		Optional<Clasz> cl = claszService.findById(code);
+		if (!cl.isPresent()) {
+			return ResponseEntity.notFound().build();
+		}
 		claszService.deleteById(code);
 		return ResponseEntity.ok().build();
 	}
